pyproject.toml
[project]
name = "oh-my-logo-cjk"
version = "0.1.0"
description = "Render text as ASCII pixel art with gradients in CLI, supporting CJK via pixel pixel-font sampling."
readme = "doc.md"
requires-python = ">=3.9"
authors = [
  { name = "oh-my-logo contributors" }
]
license = { text = "MIT" }
dependencies = [
  "typer>=0.12.3",
  "pillow>=10.3.0",
  "colorama>=0.4.6"
]

[project.scripts]
oh-my-logo-cjk = "oh_my_logo_cjk.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
managed = true

[tool.typer]
add_completion = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["oh_my_logo_cjk*"]
